Move verify.py to ../check_json.py
In preparation for actual code organization
